feat: Move widget step actions to widget folder + add Spell
