+++ src/main/java/com/questhelper/helpers/quests/coldwar/PenguinEmote.java
@@ -25,12 +25,10 @@
 package com.questhelper.helpers.quests.coldwar;
 
 import com.questhelper.questhelpers.QuestHelper;
-import com.questhelper.steps.WidgetDetails;
+import com.questhelper.steps.widget.WidgetDetails;
 import com.questhelper.steps.WidgetStep;
-import java.util.ArrayList;
 import java.util.Collections;
 import net.runelite.api.events.VarbitChanged;
-import net.runelite.client.eventbus.Subscribe;
 
 public class PenguinEmote extends WidgetStep
 {
