+++ src/main/java/com/questhelper/steps/widget/WidgetHighlight.java
@@ -22,9 +22,10 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.questhelper.steps;
+package com.questhelper.steps.widget;
 
 import com.questhelper.QuestHelperPlugin;
+import com.questhelper.steps.widget.AbstractWidgetHighlight;
 import lombok.Getter;
 import net.runelite.api.Client;
 import net.runelite.api.widgets.Widget;
